chore: Lower log level of request handler loggers to `debug` (#3064)
Closes https://github.com/electric-sql/electric/issues/3063

- Lowers `Plug.Logger`'s log level to `debug`
- Turns the query string logging to `debug`
- To offset the lack of request logs at `info` log, I've added an
additional log line for new shapes being created, so that info logs have
more continuity


This has been discussed a few times and I have the feeling that there is
agreement on this, but if anyone would like to weigh and argue for
keeping request logs in `info` please do so.

Having an HTTP service without request logs at `info` level does feel
odd so I'm not feeling 100% sure on this one.