fix: Remove shape log collector monitoring altogether (#3065)
Followup to https://github.com/electric-sql/electric/pull/3057

Simplifies connection manager as we do not need to tie
ShapeLogCollector's lifetime to the replication client - the replication
client will crash in absence of the shape log collector either way
(following @alco 's advice).

I've linked and added a special handling for the shapes supervisor exits
- if the supervisor itself exits (which it always down with `:shutdown`
which we were previously ignoring), we need to shutdown the connection
manager. This would only happen in cases of infinite restarts etc but
good to not miss this IMO.

Because the shapes supervisor is started as a temporary child to the
connection manager supervisor, if it itself dies the connection manager
wouldn't know (other than by crashing by virtue of the shape log
collector missing, in case we try to process transactions).

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>