fix: Eliminate monitor cleanup assertion flakes (#3085)
Closes https://github.com/electric-sql/electric/issues/3011

The async shape cleanup procedure is set to have a timeout of 60s, so as
to not hang on FS being slow for _too_ long. However this upper bound is
too high for tests, and at least on CI it seems that occasionally it
might get into that range (when running everything concurrently).

It's still worth investigating how it gets bottlenecked this bad, but
reducing the timeout does at least ensure that the tests are asserting
the semantics that cleanup is being called, and they don't flake because
of timeouts due to slow FS.

I was thinking of passing the timeout as a function argument but given
the module is already using `Mix.env()` to configure the code this
approach felt like the simplest, most self-contained one.


P.S. also made the slot names unique per test rather than uniform random
as some test failures seem to be related to occasional collisions of
slot names

With these changes all known (at least to me and codecov) flakes should
be covered.