feat: Feature flag as stack configuration (#3088)
In order to be able to configure feature flags per stack we need them to
not be a global config.

Trying to thread things through turned out to be a bit messy - perhaps
we should revisit the number of configuration calls we're making as I
feel we should be able to get them down to 2 (stack config and API
config I suppose?). This is ultimately what we're doing but there's many
layers of reutilization (`core_configuration` `build_shared_opts` etc)
which make it slightly confusing IMO.

One thing worth noting is that I added a default to the `Shape` schema
to be `Electric.Config.get_env(:feature_flags)` - otherwise in all tests
we'd have to explicitly specify the feature flags. I could make this
default a test only thing, let me know your thoughts.

Ultimately outside of tests, `Shape.new` is only called in one place in
the API, but in tests we call it _everywhere_.