Rename file->legacy_file and pure_file->file (#3024)
Since the introduction of pure_file, our reference documentation for
`ELECTRIC_STORAGE` has been wrong. This change aims to make it so the
`PureFileStorage` implementation is given the weight it deserves while
still making it possible to choose the legacy `FileStorage`
implementation.

No documentation changes are needed since we don't want to advertise the
use of the legacy file storage.