fix: Flush tracker acknowledgements under heavy load (#3106)
Fixes https://github.com/electric-sql/electric/issues/3102

The main issue was that whenever we starter tracking flushes for a shape
that we were not already tracking, we assumed its `last_flushed_offset`
was the `last_global_flushed_offset`. This sounds like a reasonable
assumption, and it is a valid lower bound for the new shape's last
flushed offset, but it leads to the following scenario covered by the
new test:


## Problematic Case

0. Initial state is `last_global_flushed_offset` is at transaction 9.
1. Transaction 10 for Shape A is ingested
2. Transaction 11 for Shape B is ingested
3. Flush ack for transaction 10 by Shape A is received
3a. `last_global_flushed_offset` is not updated as Shape B's most recent
flush was set to txn 9
4. Transaction 12 for Shape A is ingested
4a. We set Shape A's most recent flush as `last_global_flushed_offset`
which is txn 9, despite the fact that we had just acknowledged its flush
of transaction 10 moments ago.
5. Flush ack for transaction 11 by Shape B is received
5a. Again no updates to `last_global_flushed_offset` as Shape A's most
recent one is still at txn 9

This can go on with multiple shapes, and with enough transactions the
`last_flushed` mapping is never empty and thus
`last_global_flushed_offset` is never updated.

## Proposed Solution

The solution I have implemented is changing our assumption of what the
last flushed transaction for a newly tracked shape is. Instead of using
the `last_global_flushed_offset`, we assume the shape has flushed all
transactions up to the previous one from the one that is being ingested
now.

I claim that this is a safe assumption and a safe upper bound, because
if the shape was not being tracked when a transaction is ingested, it
means that:
1. Either the shape was previously tracked, and was flushed up to the
most recently seen transaction and removed from the mapping
2. Or the shape is new and this is the first relevant transaction
ingested for the shape

In either case, we can assume that all relevant transactions to the
shape before the current one have been flushed. If the transaction right
before this one was not relevant to the shape, and thus was not sent to
the shape, it is safe to say that it is effectively flushed, as flushing
a transaction for a shape that has nothing to write from it involves
doing nothing (it is "flushed" by virtue of it being irrelevant).


## Attempts to cover more issues

I spent some time thinking about what we could do to avoid more issues.
I tried writing a property test but it ended up not being as simple as I
was initially envisioning it.

One thing we could do is block the ingestion of new transactions if we
have not acknowledged anything for a while (or if acknowledged seen LSN
is getting significantly larger than the ackowledged flushed LSN), so as
to give time for shapes to flush and acknowledge things fully. Basically
create backpressure based on how far the seen and flushed acknowledged
LSNs are getting.

P.S. I also made a bunch of tests async as it was missing and I didn't
see any particular reason they weren't before.