fix: Avoid crashing ETS inspector on DB conn issue to preserve cache (#3112)
Overlaps a bit with @robacourt 's work, potentially relevant issue
https://github.com/electric-sql/electric/issues/3053

Minimal work to avoid having the ETS inspector crash whenever we cannot
grab a connection from the pool.

The pool dying is something to be expected throughout the lifetime of
the app, but if the pool is dead then `Postgrex.transaction` violently
exits.

In this PR I am not addressing the processes that use the inspector that
can still crash as a consequence of the failure to load stuff. The goal
of this PR is to prevent the ETS inspector itself from crashing and
taking the cache down with it, leading to more DB pool contention upon
restarting.