Fix GetExtensions to infer minimal Extensions (#3117)
Fixes the issue with the transformer which is blocking CI due to failing
typecheck.

The problem was that the `GetExtensions` type was inferring types that
aren't extensions:
```ts
GetExtensions<{
  id: number
  name: string
  org_id: number
}>
```
The snippet above was being resolved to `string | number` instead of
`never`. That was wrong because `string` and `number` are built-in types
so they don't count as extensions. Since this object type has no
extensions it should return `never`.

Fixed that by changing the `GetExtensions` type to explicitly check
whether the type extends the base `Row` type. If that's the case, then
there are no extensions and we just return `never`. Otherwise, we infer
the extensions.