chore: Simplify `PublicationManager` to only track relations (#3119)
Overdue change - simplifies `PublicationManager`.

No more complicated multi-index ref counting - just relations.

I've moved most things to use mapsets as they are appropriate for the
set of relations we are tracking - we were doing lots of `Enum.uniq`
calls in various places anyway. Not that the number of relations should
ever be so large that this becomes a concern.