feat(sync-service): Improve performance of LRU shape expiry (#3141)
TurboLearn.ai have been seeing 2-8s latency on requests when the number
of shapes exceeds the maximum number of shapes and LRU expiry kicks in.
This is due to the shape cache process being blocked while the expiry
happens, which can take a few seconds.

This PR moves expiry out of the ShapeCache process and into a new
ExpiryManager process. When a new shape is added, the ShapeCache calls
`ExpiryManager.notify_new_shape_added/1` which is non-blocking and
debounced, and if max_shapes has been set, triggers finding the LRU
shapes and expiring them.