docs: sync stacks page. (#3138)
Was starting on the tutorial and went back and checked the previous
"Sync busters!" concept in
https://github.com/electric-sql/electric/pull/2584 in case it had
anything useful to keep / port over. Played with it and ended up
converting into this "Sync stacks" page.

The idea is to have a page that emphasises "composable" + "recommend
TanStack DB" and has the common content we often relay about choice of
stack / options / what's best for app dev, how do TanStack DB and PGlite
relate etc.

=> https://deploy-preview-3138--electric-next.netlify.app/docs/stacks

I would envisage publishing alongside the new Quickstart and Tutorial,
below the Tutorial in the left hand nav. Sharing early / separately
because (a) it's largely self-contained and (b) for feedback around
whether this page adds value/clarity or reduces it.