Stop externalizing ShapeStatus' state and spreading it over the codebase
Generally speaking, this is unconventional, even within our own
codebase. ShapeStatus should be responsible for naming its own ETS
tables and encapsulating any other state pertaining to its function.

Next, it doesn't need to keep storage in its state because it's only
used for saving/restoring from backup. During initialisation, storage is
passed to it as an argument. Afterwards, only the backup directory path
is kept in the ShapeStatusOwner's state to be used at termination time.

Finally, the way Consumer tests are written made it extra difficult to
forego dynamic setting of the shape status module, so I left this option
for use in tests.