feat: Simplify purging of shapes in case of timeline or slot change (#3146)
This is in response to slow purge observed by @alco 

- In case of timeline change or slot change, simply delete the entire
shape directory
- If publication manager refresh fails during shape cache init,
synchronously wait for all shapes to be cleaned up rather than async
purging
- Do not do early "deregistrations" of shapes from shape status as it
might mess with operations happening on the consumer. The only ones
deregistering the shape should be 1) the consumer dying or 2) the
cleanup task supervisor that asserts the consumer is dead before doing
so. (@magnetised please confirm cause you added the early deregistration
logic)