feat: Decouple `PublicationManager` startup from `ShapeCache` startup (#3154)
Continuing with simplifying startup/cleanup procedure.

Before, we were "recovering" the publication manager state in memory
before updating the publication, and if that failed we cleaned up all
the shapes and failed shape cache initialisation.

This coupling was hard to reason about in the entirety of the system,
and we are moving towards more aggressive, quicker full shape cleanups
that will be done outside of shape cache (directory cleanup).

Already this method wasn't actually doing what it was supposed to
(reducing number of publication updates), as some shapes that were being
recovered would not have had a snapshot and thus would've triggered an
`add_shape` and thus publication updates.

The publication manager is perfectly capable of handling multiple
requests to add shapes to the publication quickly and efficiently, so
I've made it always call `add_shape` regardless of whether the snapshot
exists. This allows us to remove the `recover_shape` method entirely.

Ideally I would want to move out the publication adding fro the
snapshotter and put it in the consumer initialisation, but we are
relying on its failures and parse them into `SnapshotError` at the
moment so I will do that in a separate PR.