Extract shape's last_read timestamp into a separate ETS table (#3151)
Listing least recently used shape handles becomes more efficient.

Counting shapes becomes an O(1) operation.

The the shape counting operation being cheap now, we can remove the
complexity of spamming ExpiryManager's inbox and debouncing expiry
checks.

If the tenant is in a steady state where it doesn't see any new shape
creations, the cost of the periodic check is a call to `:ets.info(...,
:size)`. On the other end of the spectrum, if the tenant is constantly
churning through new shapes, sending a message for each shape to
ExpiryManager is pure overhead.

The current debouncing implementation violates OTP's conventions by
using Process.sleep() inside a GenServer callback. This results in the
ExpiryManager process accumulating messages in its inbox which
immediately stands out among other processes in the Observer or
observer_cli. A periodic check is simpler and is less taxing on the VM.