fix: reduce file handle usage to 1 on PFS (#3098)
Closes #3078

This PR lowers the amount of needed constantly open file handles to 1
(and that's also closed on shape inactivity after 30 seconds by default
as per #3073)

Does 2 things:
1. The chunk file is opened only when chunk boundaries are written -
that's once every 10MB of JSON data. If it's opened once within a
transaction, it's not closed until the transaction is fully written, to
avoid additional overhead if the txn is large or spans a boundary (and
we need to close current and open a new chunk).
2. Foregoes opening the key file completely. Instead we're paying that
cost at compaction time by building the key from the file. This makes
the compaction more expensive, but we're keeping around the keys for
compacted data so that work is not re-done, and compaction is a
background process anyway. We can roll this change back if need be.

---------

Co-authored-by: msfstef <msfstef@gmail.com>