feat: Faster large cleanups using renames and background batch deletes (#3158)
For large numbers of shapes, deleting all of them takes a long while
(doing an Elixir/Erlang rm_rf fo 10k shapes is in the order of tens of
seconds) - `rm_rf`ing recursively lists into directories and deletes
things which for large directories like deleting all shapes or all files
for a shape can be slow.

Using a different `rm_rf` implementation like the regular util is faster
(twice as fast) but messes with the BEAM and we would have to use
ports/nifs/dirty schedulers which I want to avoid for the time being for
the purpose of keeping things as portable and embeddable as possible.
Also Erlang's `rm_rf` implementation is safe and should play nice with
schedulers if done in the background without blocking our critical
processes.

I've implemented an `AsyncDeleter` process that lives in the top level
of our stack - all it does is it maintains a temporary trash directory
(stack scoped), it has a public API called `delete` which all it does
is:
1. Ensure trash directory exists with `mkdir_p!` to guarantee move will
work
2. Attempt to rename desired path to delete into trash directory (with
random suffix to guarantee uniqueness)
2a. If file already does not exist when attempting to rename it, ignore
it and move on
3. If rename succeeds, cast a message to `AsyncDeleter` to schedule a
cleanup
4. `AsyncDeleter` performs a cleanup of the entire stack trash directory
every `cleanup_interval_ms`

I've added some tests, including one to ensure concurrent deletes and
cleanups work, since the renames happen in the calling process, and the
cleanup can occur at any time.

For now I've only made `cleanup!` and `cleanup_all!` storage methods use
the async deleter, which are the main ones that are slow, but we can
potentially use it for more deletes in our system (compaction does a lot
of deletes for example).