feat: add timeout to first data on snapshots to avoid long-running queries (#3156)
Adds a 30s timeout to first data batch - otherwise query is considered
too slow for snapshots.

Timeout is counted only after we've made a connection and started a
transaction - pool queueing and other waits are not included. We're
waiting until first data (or end-of-data in case of absence) for those
30s for the query to be considered "good enough". Probably can make that
timer even lower. The query itself is still unbounded - if the snapshot
is multiple GBs that are being copied, we'll not interrupt it in this
PR.

Closes #3157