fix(examples/linearlite): fix loading if issue url is opene with no synced issue database (#2238)
I was trying to debug something unrelated and found that the IssuePage
was getting into a bad state if you browsed directly to it in an
incognito tab without visiting the main page first.

The fix is reasonably trivial: delay all rendering queries until the
initial db dump in synced, like how the rest of the example works.

I also fixed the "not found" rendering. I don't think that
`useLiveQuery(liveIssue).rows[0]` can ever be `null`: it will always be
`Item | undefined` (typescript types it as `Item` unless you turn on
noUncheckedIndexedAccess).

# steps to test

* check that navigating to issue with no synced issue database works
* open http://localhost:5173/ and click on an issue to get issue url (it
will look like
http://localhost:5173/issue/d8af89e7-d1cf-4488-8b1b-424f120d0aa6)
  * open a new incognito window (⌘⇧N) and go to that url
* in main, you would get `Loading...` forever (it would fix itself if
you refresh the page)
    * on this branch, it should load correctly
* check that navigating to a nonexistent issue tells you "not found"
* open an issue that doesn't exist (e.g.
http://localhost:5173/issue/00000000-0000-0000-0000-000000000000)
    * on main, you would get `Loading...`
    * on this branch, you should get `Issue not found`