feat(sync-service) Optimise removing a shape from the where clause filter indexes (#3160)
Turbo.ai are seeing ShapeLogCollector.unsubscribe take 100ms on average
with a P99 of 300ms. This blocks the replication stream and given that
50 shapes can be expired at a time can cause many seconds of lag.

We've seen from telemetry that it's removing from the where clause
filter indexes that is the bottleneck.

This PR speeds up removing a shape from the where clause filter indexes.
For 60K shapes, it speeds it up 12800x. It achieves this by using the
shape data to selectively remove from the index tree rather than
traversing the whole tree looking for the shape handles.

This PR also lowers the memory footprint of the filter. A naive
implementation would have increased the memory footprint since the shape
data is needed for removal now, but since we're trying to lower memory
use I've taken this opportunity to remove the duplication of shape data
in the filter so now the memory use is less than before.


```
# before
Removing 100 shapes from 60000: each took 12.805ms mean, P99 48.303ms

:erts_debug.flat_size(filter) #=> 12581546

# this PR
Removing 100 shapes from 60000: each took 0.001ms mean, P99 0.008ms

:erts_debug.flat_size(filter) #=> 12324026
```