fix: Handle errors caused by stack termination
At any point in the API request cycle the referenced stack could go
down: either due to some problem with the database connection or because
of some cloud restart.

We should handle errors due to the stack disappearing without obscuring
errors arising from other issues/bugs in the request handling.

This PR wraps the API.validate/2 and API.serve_shape_log/1 calls in
exception handling and a re-test of the stacks availability. If the
stack is down then we return the correct 503 response. If its up then we
retry the call on the understanding that if the stack was the problem
it's now up again and things should work but if the problem was
somewhere else the same request will raise the same exception which we
won't catch/retry.