fix: Prevent race between shape adding and removing (#3174)
Fixes https://github.com/electric-sql/electric/issues/3173

Our `remove_shape` call uses a `select_delete` to remove all relevant
records from the table for a given shape.

Because `select_delete` does a traversal and does not guarantee that
concurrent operations won't see mixed results during it (i.e. it is not
atomic and isolated), we ran into the following scenario:

1. We check `get_existing_shape` to determine whether a handle exists
for a given shape spec
2. That checks the `shape -> handle` lookup, which returned a handle
3. That _subsequently_ checks the `handle -> metadata` lookup for the
latest offset on the shape, which returns nil
4. The overall `get_existing_shape` call returns nil (assuming that if
anything is missing, all records are missing)
5. We get to `add_shape` which expects `:ets.insert_new` to succeed for
the entry to the `shape -> handle` lookup
6. The operation fails as the `shape -> handle` lookup entry is still
present (see step 2)

I've opted for the following solution:
- In `remove_shape`, ensure the `shape -> handle` lookup entry gets
deleted _first_, as it is the pointer from a shape def to any existing
shapes.
- The shape cleanup can be delayed (similarly for the files as well),
but the first thing that needs to happen is to "deprecate" it such that
the shape spec is no longer pointing to its handle
- In `get_existing_shape`, rather than swallowing the error in case the
`shape -> handle` entry is present, but the `handle -> metadata` entry
is not, I'm asserting that it _should_ be present.
- The swallowing of the error in that call was essentially what was
causing this race silently
- An alternative solution would have been for `get_existing_shape` to
call itself recursively until it returns a consistent result, but I
thought that would be disgusting.

## NOTES
I've managed to reproduce this in a test, but I needed to actually seed
the shape status table with lots of shapes such that the traversal made
by `select_delete` _actually_ takes some time and produces these races.
How often it would fail obviously depends on the machine, concurrency
tests can often produce false positives, but it's better than nothing.

We need separate test suites for high concurrency scenarios as they take
time to build up a seeded scenario and run for a longer time. But they
are the only ones that test things at scale.