docs: replace LLMs guide and llms.txt with AGENTS.md. (#3148)
See:

- https://deploy-preview-3148--electric-next.netlify.app/docs/agents
- https://deploy-preview-3148--electric-next.netlify.app/AGENTS.md

Note that it's setup to expect the new quickstart, so let's merge with
that. That aside, it's working pretty well for me testing in claude,
codex and gemini. Take it for a spin and see if works for you / if you
can find issues with it.

N.b.: re: length
https://chatgpt.com/s/t_68cb5bd770bc81919a499da1ae32c31c

---------

Co-authored-by: Kyle Mathews <mathews.kyle@gmail.com>