fix: Handle connection not available error from inspector correctly (#3177)
Fixes https://github.com/electric-sql/electric/issues/3175

I'm not super familiar with Ecto so I tried to stay as close to the end
of the pipeline as possible for this.

I don't love that the inspector returns an `:ok` tuple _or_ a
`:table_not_found` _or_ and `:error` tuple - either the table not found
error should be in the error tuple or `connection_not_available` should
not be in one - but we can fix with a subsequent PR.

Happy to accept feedback or direct fixes on this PR, just want to fix
this error asap.