Partition DynamicConsumerSupervisor to improve startup and shutdown times (#3183)
This adds a supervisor layer where a `PartitionSupervisor` now carries
the `DynamicConsumerSupervisor.name(stack_id)` name and it starts a
number of `DynamicConsumerSupervisor` processes to shard load between
them. This change is transparent for the rest of the code that uses
`DynamicConsumerSupervisor`'s public API.

I've done some measurements by repeatedly starting Electric, waiting for
it to start all consumer processes for 50K shapes, then terminating it
with the `SIGTERM` signal. I have done 5 attempts for the current
implementation in `main` and for this PR, measuring both the startup and
shutdown time.

|  | Attempt 1 | Attempt 2 | Attempt 3 | Attempt 4 | Attempt 5 |
| --- | --- | --- | --- | --- | --- |
| PartitionSupervisor startup | 28.8s | 12.5s | 11.8s | 11.0s | 11.8s |
| DynamicSupervisor startup | 20.6s | 20.0s | 20.0s | 19.7s | 20.5s |

|  | Attempt 1 | Attempt 2 | Attempt 3 | Attempt 4 |
| --- | --- | --- | --- | --- |
| PartitionSupervisor shutdown | 4.8s | 10.1s | 11.0s | 11.0s |
| DynamicSupervisor shutdown | 27.0s | 27.4s | 28.2s | 28.6s |

The 1st attempt can be considered an outlier since timings have become
consistent afterwards.

When a PartitionSupervisor is used, all processes shut down within 2-3
seconds and the rest of the shutdown time is dominated by the shape
status table backup.