feat: add a control message at the end of snapshot (#3163)
@samwillis requested this as an explicit boundary so that it's present
in the replication stream and allows clients to reason about
transactions that have been incorporated into the initial snapshot,
which is important for write-then-read pattern in Electric, where we're
waiting for the txid returned from a read.

This add a new control message - `snapshot-end`, with 3 additional
fields: xmin/xmax/xip which should be plenty.