fix(sync-service): Add stack canary process to detect shutdown
The last process in a supervisor's children list will be the first to be
terminated in a restart. This canary process registers itself with the
stack status and any shutdown of the Recplication.Supervisor, either due
to a crash or a stack restart, will mark the stack as down immediately.

Fixes #3190