fix: Only register the writer consumer when terminating (#3199)
Currently we register all shape consumer processes eagerly with the
monitor and then use this to perform a cleanup when they terminate with
the reason `{:shutdown, :cleanup}`.

In reality we don't need to register all the writers, only the ones who
are requesting a cleanup. So rather than register in the `init/1`
callback, this pr merges the writer registration with the request for
reader termination notification.

With this a supervisor kill all -> restart won't even touch the monitor
but shapes that have requested a clean termination will be properly
handled.

Fixes #3191