fix(sync-service): Add stack canary process to detect shutdown (#3203)
The last process in a supervisor's children list will be the first to be
terminated in a restart. This canary process registers itself with the
stack status and any shutdown of the Recplication.Supervisor, either due
to a crash or a stack restart, will mark the stack as down immediately.

Seems to prevent a lot of errors on stack termination locally: only
active requests raise due to the exit of the "crashed" process rather
than every request after the process crash until the consumers have been
terminated.

Fixes #3190