fix: Validate connection options for lock breaker connection (#3236)
_sort of_ fixes [this
error](https://electricsql-04.sentry.io/issues/66229550/?environment=production&project=4508410462404688&query=is%3Aunresolved&referrer=issue-stream)
- we should be using validated connection options for the lock breaker
as well

More generally, since we use `sync_connect: true` for the lock breaker
conn, we fail completely if we don't manage to connect which can happen
for a variety of reasons.

I've converted this to a `with` clause so that we only try and break the
lock if we can start the connection, but ignore the failure otherwise
with a simple warning.