Stop reporting disk usage in StackTelemetry (#3229)
The metric is measured by recursively obtaining the size of every single
file in the stack's on-disk storage. For stacks that have tens of
thousands of shapes, a single measurements takes tens of seconds. This
is clearly not suitable for a metric that is polled every 5 seconds.

We could keep a running size of each shape's on-disk log data if needed.
But it seems more appropriate to obtain the disk usage numbers outside
of Electric, since the organization of stacks and shapes on disk is
straightforward enough to lend itself to grouping and slicing without
requiring any specific metadata from Electric itself.