fix: Ensure pool process exit does not log unnecessary error (#3265)
Fixes [this
issue](https://electricsql-04.sentry.io/issues/47739402/?environment=production&project=4508410462404688&query=is%3Aunresolved&referrer=issue-stream)
where the pool exiting logs an error because we are not using a
`{:shutdown, error}` tuple.

I've made it exit with an expected shutdown and we strip the reason from
the tuple in the manager instead.

I've taken the opportunity to also parse [this Postgrex
error](https://electricsql-04.sentry.io/issues/60380806/?environment=production&project=4508410462404688&query=is%3Aunresolved&referrer=issue-stream)
about not being able to connect to a compute node as retryable.