fix: Ensure abort error is emitted if fetch aborted while consuming response body (#3266)
This fixes the TS client flakes - if the request is aborted via the
abort controller while consuming the response body, we should emit the
same abort error that we do throughout the rest of the lifetime of the
request, such that it is handled correctly by the stream.

Also removed unnecessary assertion in Elixir sync service test which I
had left there in previous flaky test fix.