feat: Publication manager updates affect only relevant shapes (#3218)
Fixes https://github.com/electric-sql/electric/issues/3186

This PR includes significant changes to
`Electric.Postgres.Configuration` and
`Electric.Replication.PublicationManager`.


### Changes
- Ensure PubMan handles per-relation errors such that only relevant
shapes fail
- If any individual relation fails to be configured, relevant shapes are
terminated
- If global error (e.g. connection error) occurs, propagate error to all
requesters but do not invalidate shapes explicitly and keep process
alive and able to respond (see work on resilience to network failures
@robacourt )
- Got rid of retries - we don't do them everywhere, e.g. snapshots, and
should be handled at network layer (see @robacourt 's work)
- Configuration now returns map with configuration result for each
relation it either touched or detected to be invalid
- Validating the configuration returns the same result for simplicity of
API, with different error codes since it cannot actually modify the
relations
- PubMan now also checks that the publication 1) exists, 2) is
configured to publish all operations, and 3) is owned by the Electric PG
user.
- These are all checks done once at the start of setting up the
replication connection and publication, but are replicated in PubMan
which will be doing these before every publication update/check, which
also occur periodically, so we can detect something going wrong
- These checks will also set the `can_alter_publication?` flag if the
ownership changes so appropriate results are returned for shape requests
- Only log errors for unknown errors when failing to configure a
relation, everything else is a warning
- _Slightly_ more generous with calling `ALTER TABLE` and `ALTER
PUBLICATION` when fixing misconfigured relations just for the sake of
API simplicity - no reason to be so cheap for something that doesn't
happen often at the expense of clarity IMO.


## TODOs
- If the publication does not exist, we should notify the connection
manager so that the whole stack is restarted. We are not doing this
right now, we are just shutting down the publication manager which just
restarts the replication supervisor but doesn't actually fix the
problem. I'll add this as a separate PR.
- Cover case where we don't have ownership over the publication, but we
do over the table, so we would need to set the replica identity without
trying to add it to the publication.